+++ python/openassetio/pluginSystem/PluginSystemManagerImplementationFactory.py
@@ -65,8 +65,8 @@ class PluginSystemManagerImplementationFactory(ManagerImplementationFactoryInter
             self.__paths = os.environ.get(self.kPluginEnvVar, "")
             if not self.__paths:
                 self._logger.log(
-                    ("%s is not set. It is somewhat unlikely that you will "
-                     + "find any plugins...") % self.kPluginEnvVar, self._logger.kWarning)
+                    self._logger.kWarning, ("%s is not set. It is somewhat unlikely that you will "
+                     + "find any plugins...") % self.kPluginEnvVar)
 
         self.__pluginManager = PluginSystem(self._logger)
 
@@ -100,7 +100,7 @@ class PluginSystemManagerImplementationFactory(ManagerImplementationFactoryInter
         if not self.__pluginManager:
             self.__scan()
 
-        self._logger.log(f"Instantiating {identifier}", self._logger.kDebug)
+        self._logger.log(self._logger.kDebug, f"Instantiating {identifier}")
         plugin = self.__pluginManager.plugin(identifier)
         interface = plugin.interface()